Benchmark SQLite for large blobs against file based I/O
Motivated by this article [1] I wanted to know if SQLite might be
faster, or at least almost as fast, as file I/O for blobs that range
from 1MB to about 100MB.  Good performance with those blob sizes would
make SQLite suitable as a simple file storage, e.g. for a simple
self-contained webservice where a backup would only require to
download the sqlite database (I hate maintenance work).

Running the benchmarks showed that reading big blobs from SQLite is about
5 times slower than reading them from disk.  Note that the benchmarks
  were run on a blazing fast Samsung 970 evo nvme SSD, so your mileage
  will vary.

Also, a major downside when using the most common driver (`mattn/go-sqlite3`)
is that the blob must be buffered completely in memory when INSERTing it
into the database.  This downside can be prevented when using
crawshaw/sqlite which is not a `database/sql` driver and thus needs some
work to get used with its API.  But, it provides streaming I/O for
sqlite which is nice.  Unfortunately, I could not see any performance
benefit from using it.

You may wonder why there are two `bench.go` files, the reason is that
compiling will fail when mattn/go-sqlite3 and crawshaw/sqlite are
present because then the SQLite symbols are present twice.  This might
be solved by using build tags but I went for the pragmatic solution and
went with two files.

[1]: https://www.sqlite.org/fasterthanfs.html