Try to parse OSC addresses and address patterns
Disclaimer:  This will not work, do not use this for any serious OSC
address parsing!

The specification is not very specific and does not go into detail about
the combination of pattern expression, e.g. can '*' follow a range and
and so on [1].  Building my own matcher proves to be a complicated task
and existing implementations just translate the patterns into regex [2]
because of the specifications ambiguity.

Anyways, learning how to use nom was very nice and will certainly be
helpful in the future.

[1]: http://opensoundcontrol.org/spec-1_0.html#osc-message-dispatching-and-pattern-matching
[2]: https://github.com/attwad/python-osc/blob/master/pythonosc/dispatcher.py